adaptación merchandising en el proyecto

+++ src/main/java/org/springframework/samples/petclinic/service/MerchandasingService.java
@@ -1,11 +1,16 @@
 package org.springframework.samples.petclinic.service;
 
+import java.util.ArrayList;
+import java.util.Collection;
 import java.util.List;
+import java.util.Optional;
+
+import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataAccessException;
 import org.springframework.samples.petclinic.model.Merchandasing;
-
+import org.springframework.samples.petclinic.model.TipoMerchandasing;
 import org.springframework.samples.petclinic.repository.MerchandasingRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -32,7 +37,7 @@ public class MerchandasingService {
 //	}
 
 	@Transactional
-	public void saveMerchandasing(Merchandasing merchandasing) throws DataAccessException {
+	public void saveMerchandasing(@Valid Merchandasing merchandasing) throws DataAccessException {
 		merchandasingRepository.save(merchandasing);
 
 	}
@@ -43,14 +48,47 @@ public class MerchandasingService {
 	}
 
 	@Transactional(readOnly = true)
-	public Merchandasing findMerchandasingById(int id) throws DataAccessException {
-		return merchandasingRepository.findById(id);
+	public Merchandasing findMerchandasingById(int id) throws IllegalArgumentException {
+		Optional<Merchandasing> merch = this.merchandasingRepository.findById(id);
+		
+		if (merch.isPresent()) {
+			return merch.get();
+		} else {
+			throw new IllegalArgumentException("El merchandising no existe");
+
+		}
+
 	}
 
 	@Transactional
 	public void deleteMerchandasing(int merchandasingId) throws DataAccessException {
-		merchandasingRepository.deleteById(merchandasingId);
+		Optional<Merchandasing> merch = this.merchandasingRepository.findById(merchandasingId);
+		
+		if (merch.isPresent()) {
+			merchandasingRepository.deleteById(merchandasingId);
+		}else {
+			throw new IllegalArgumentException("El merchandising que desea borrar no existe");
+		}
+		
+		
 
+	}
+
+	public void delete(Merchandasing m) {
+		Optional<Merchandasing> merch = this.merchandasingRepository.findById(m.getId());
+		if (merch.isPresent()) {
+			merchandasingRepository.delete(m);
+		}else {
+			throw new IllegalArgumentException("El merchandising que desea borrar no existe");
+		}
+	}
+
+	public Collection<TipoMerchandasing> getTipos() {
+		Collection<TipoMerchandasing> tipos = new ArrayList<TipoMerchandasing>();
+		for (TipoMerchandasing t : TipoMerchandasing.values()) {
+			tipos.add(t);
+		}
+		return tipos;
 	}
 
 }
